Add files via upload
This graph depicts the distribution of the sentiment of tweets produced by Donald Trump and Hillary Clinton. The red boxplot shows the distribution of the favorite count of tweets with a negative sentiment, while the blue boxplot shows the distribution of the favorite count of tweets with a positive sentiment.